fix: KNX-DPT-Skalierung HSV, Dimmzeiten (uint32) und KO-Null-Guards
- RGB: HSV-Einzelkanal-KOs skalieren jetzt auf 0-255 vor hsvToRGB
  (H aus DPT_Angle *255/360, S/V aus DPT_Scaling *2.55) - vorher wurden
  KNX-Engineering-Werte (0-360 / 0-100) ungewandelt an eine Funktion
  gegeben, die 0-255 erwartet -> falsche Farben. Konsistent zum
  bestehenden *2.55 im R/G/B-Pfad; belegt gegen dptconvert.cpp.
- Dimmzeiten: m_durationrelativ/absolut von uint16_t auf uint32_t
  (getTimeWithPattern liefert ms bis in den Millionenbereich; eine
  2-min-Zeit wurde bisher schon bei der Zuweisung auf ~54 s gekuerzt).
  (word)-Cast auf _time entfernt (sekundaere Kuerzung). EK/TW/RGB.
- LEDHelper::getTimeWithPattern rechnet in uint32_t (kein Signed-Int-
  Overflow mehr bei Minuten/Stunden).
- _time/_lastTaskExecution in EK/TW/RGB mit 0 initialisiert.
- LEDModule::processInputKo: Null-Guards vor channelEK/TW/RGB[i], damit
  ein KO auf einen in der Betriebsart unbelegten Slot nicht crasht.

+++ src/LEDHelper.cpp
@@ -190,13 +190,13 @@ uint32_t LEDHelper::getTimeWithPattern(uint16_t time, uint8_t base)
     switch (base)
     {
     case TIMEBASE_TENTH_SECONDS:
-        return time * 100;
+        return (uint32_t)time * 100UL;
     case TIMEBASE_SECONDS:
-        return time * 1000;
+        return (uint32_t)time * 1000UL;
     case TIMEBASE_MINUTES:
-        return time * 60000;
+        return (uint32_t)time * 60000UL;
     case TIMEBASE_HOURS:
-        return time * 3600000;
+        return (uint32_t)time * 3600000UL;
     default:
         return 0;
     }
